refactor for code cleaning
